requirements.txt
flask
python-dotenv
azure-cognitiveservices-vision-computervision
google
beautifulsoup4
googlesearch